user register done

+++ data/response/users.go
@@ -1,7 +1,7 @@
 package response
 
 type UserRegRes struct {
-	Age			uint	`json:"age"`
+	Age			uint8	`json:"age"`
 	Email		string	`json:"email"`
 	Id			uint	`json:"id"`
 	Username	string	`json:"username"`
